Added device tree support for address and irq.
Until now the interrupt numbers and addresses for osif and proc control
kernel modules were hard coded. Due to recent changed in the linux
kernel we need to configure the interrupts in the device tree to be
initialized correctly. Furthermore, we need to use the device tree
parsing to get the correct interrupt numbers.
To allow usage for newer kernels we add support for the device tree with
this commit. Note, that this support is only a hot fix and not a
permanent solution which we are currently working on in the develop_ic
branch where we use the correct way of loading the kernel module for a
device tree entry. However, for normal usage this should also work fine.

From now on you need to add the follwing ReconOS components to the
device tree:

	amba: amba {
		reconos_osif: reconos_osif@75a00000 {
			compatible = "upb,reconos-osif-3.1";
			reg = <0x75a00000 0x10000>;
		};

		reconos_osif_intc: reconos_osif_intc@7b400000 {
			compatible = "upb,reconos-osif-intc-3.1";
			reg = <0x7b400000 0x10000>;
			interrup-parent = <&intc>;
			interrupts = <0 58 4>;
		};

		reconos_proc_control: reconos_proc_control@6fe00000 {
			compatible = "upb,reconos-proc-control-3.1";
			reg = <0x6fe00000 0x10000>;
			interrupt-parent = <&intc>;
			interrupts = <0 59 4>;
		};
	};